week3/synthetic_data_generator/requirements.txt
gradio>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
transformers>=4.41.0
torch>=2.2.0
accelerate>=0.30.0
sentencepiece>=0.2.0
